requirements/main.txt
aiocsv==1.3.2
aiofiles==24.1.0
aiosqlite==0.21.0
asyncer==0.0.8
graphviz==0.20.3
httpx<1
jupytext
nest_asyncio==1.6.0
numpy<=2.2.5
pandas>=2
parso==0.8.4
pyautogen[openai]==0.8.7
pydantic>=2.10.2,<3
typer>=0.9,<0.16
